tasks/requirements.txt
fastapi
uvicorn[standard]
sqlmodel[asyncio]
asyncpg
httpx
pydantic-settings
redis
python-dateutil
# common_lib is installed manually in CI or local dev environment
# common_lib @ file://localhost/common